add a slideChooser
the user can specify a strategy for which slide to choose. this strategy can vary from "use this specific slide" to "use a random one" to the default "use the first one". This is made more complicated by gslides not using sequentially numbered slide ids per deck. This change adds a function that can take the user's strategy or a default and return a function suitable for use in the promise chain