restructures exports
Corsica expects the imported library to present a function, and node's require is dumb as you can imagine. The 0.5.0 version returns an object that needs to be destructured. This shift was made to enable testing of the component functions.  Corsica doesn't rely on es6 modules, though, so simply naming one of those functions to be the `default` isn't sufficient. Instead this patch moves most of the business logic over to an impl function that can be tests, and then imports and re-exports those through a facade that matches Corsica's expectation